requirements.txt
# requirements.txt
torch>=2.1.0
tensorflow-macos>=2.14.0
tensorflow-metal>=1.1.0
numpy>=1.24.3
whisper-openai>=1.1.10
spacy>=3.7.2
transformers>=4.35.0
sounddevice>=0.4.6
pyaudio>=0.2.13
librosa>=0.10.1
